split quadratic
